Create Ant build files.
Builds with the same source layout as the Leiningen build, but
different dest.  The Ant script contains a 'one-jar' build target that
will build a jar file of jarfiles of the dependencies, and includes
the one-jar bootloader that is capable of loading from these nested
Jarfiles.  The author of one-jar says that including the license into
the resulting jarfile is sufficient to comply with the author's
requests.

The one-jar archives take a bit longer to startup than "uber-jars"
that Maven and Leiningen make.